# docs/design/pre-registration/service/Booking-Service-API-Spec.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'integ.mosip.io:443'
basePath: /booking
tags:
  - name: Booking
    description: Booking Controller
paths:
  /v0.1/pre-registration/booking/appointmentDetails:
    get:
      tags:
        - Booking
      summary: Fecth Appointment details
      operationId: appointmentDetailsUsingGET
      produces:
        - application/json
      parameters:
        - name: pre_registration_id
          in: query
          description: pre_registration_id
          required: true
          type: string
      responses:
        '200':
          description: Appointment Booked Successfully
          schema:
            $ref: '#/definitions/MainResponseDTO«BookingRegistrationDTO»'
        '400':
          description: Unable to Book the appointment
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /v0.1/pre-registration/booking/availability:
    get:
      tags:
        - Booking
      summary: Fetch availability Data
      operationId: getAvailabilityUsingGET
      produces:
        - application/json
      parameters:
        - name: registration_center_id
          in: query
          description: registration_center_id
          required: true
          type: string
      responses:
        '200':
          description: Availablity details fetched successfully
          schema:
            $ref: '#/definitions/MainResponseDTO«AvailabilityDto»'
        '400':
          description: Unable to fetch the records
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /v0.1/pre-registration/booking/book:
    post:
      tags:
        - Booking
      summary: Booking Appointment
      operationId: bookAppoinmentUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: bookingDTO
          description: bookingDTO
          required: true
          schema:
            $ref: '#/definitions/MainListRequestDTO«BookingRequestDTO»'
      responses:
        '200':
          description: Appointment Booked Successfully
          schema:
            $ref: '#/definitions/MainResponseDTO«List«BookingStatusDTO»»'
        '201':
          description: Created
        '400':
          description: Unable to Book the appointment
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - Booking
      summary: Cancel an booked appointment
      operationId: cancelBookUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: requestDTO
          description: requestDTO
          required: true
          schema:
            $ref: '#/definitions/MainRequestDTO«CancelBookingDTO»'
      responses:
        '200':
          description: Appointment canceled successfully
          schema:
            $ref: '#/definitions/MainResponseDTO«CancelBookingResponseDTO»'
        '201':
          description: Created
        '400':
          description: Unable to cancel the appointment
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /v0.1/pre-registration/booking/bookedPreIdsByRegId:
    post:
      tags:
        - Booking
      summary: Reterive all pre-registration ids by registration center id
      operationId: getPreIdsByRegCenterIdUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: requestDTO
          description: requestDTO
          required: true
          schema:
            $ref: '#/definitions/MainRequestDTO«PreRegIdsByRegCenterIdDTO»'
      responses:
        '200':
          description: Reterived all pre-registration ids successfully
          schema:
            $ref: >-
              #/definitions/MainListResponseDTO«PreRegIdsByRegCenterIdResponseDTO»
        '201':
          description: Created
        '400':
          description: Unable to reterive pre-registration ids
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /v0.1/pre-registration/booking/masterSynchronization:
    get:
      tags:
        - Booking
      summary: Sync master Data
      operationId: saveAvailabilityUsingGET
      produces:
        - application/json
      responses:
        '200':
          description: Master Data Sync is successful
          schema:
            $ref: '#/definitions/MainResponseDTO«string»'
        '400':
          description: Unable to fetch the records
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
definitions:
  AvailabilityDto:
    type: object
    properties:
      centerDetails:
        type: array
        items:
          $ref: '#/definitions/DateTimeDto'
      regCenterId:
        type: string
    title: AvailabilityDto
  BookingRegistrationDTO:
    type: object
    properties:
      registration_center_id:
        type: string
        description: Registration Center ID
      appointment_date:
        type: string
        description: Appointment Date
      time_slot_from:
        type: string
        description: From Time Slot
      time_slot_to:
        type: string
        description: To Time Slot
    title: BookingRegistrationDTO
  BookingRequestDTO:
    type: object
    properties:
      preRegistrationId:
        type: string
        description: Pre-Registration ID
      oldBookingDetails:
        description: Old Booking Data
        $ref: '#/definitions/BookingRegistrationDTO'
      newBookingDetails:
        description: New Booking Data
        $ref: '#/definitions/BookingRegistrationDTO'
    title: BookingRequestDTO
  BookingStatusDTO:
    type: object
    properties:
      bookingMessage:
        type: string
      bookingStatus:
        type: string
      preRegistrationId:
        type: string
    title: BookingStatusDTO
  CancelBookingDTO:
    type: object
    properties:
      pre_registration_id:
        type: string
        description: Pre-Registartion ID
      registration_center_id:
        type: string
        description: Registration Center ID
      appointment_date:
        type: string
        description: Appointment Date
      time_slot_from:
        type: string
        description: From Time Slot
      time_slot_to:
        type: string
        description: To Time Slot
    title: CancelBookingDTO
  CancelBookingResponseDTO:
    type: object
    properties:
      message:
        type: string
      transactionId:
        type: string
    title: CancelBookingResponseDTO
  DateTimeDto:
    type: object
    properties:
      date:
        type: string
      holiday:
        type: boolean
      timeSlots:
        type: array
        items:
          $ref: '#/definitions/SlotDto'
    title: DateTimeDto
  ExceptionJSONInfoDTO:
    type: object
    properties:
      errorCode:
        type: string
      message:
        type: string
    title: ExceptionJSONInfoDTO
  LocalTime:
    type: object
    properties:
      hour:
        type: integer
        format: int32
      minute:
        type: integer
        format: int32
      nano:
        type: integer
        format: int32
      second:
        type: integer
        format: int32
    title: LocalTime
  MainListRequestDTO«BookingRequestDTO»:
    type: object
    properties:
      id:
        type: string
        description: request id
      ver:
        type: string
        description: request version
      reqTime:
        type: string
        format: date-time
        description: request time
      request:
        type: array
        description: list of request
        items:
          $ref: '#/definitions/BookingRequestDTO'
    title: MainListRequestDTO«BookingRequestDTO»
  MainListResponseDTO«PreRegIdsByRegCenterIdResponseDTO»:
    type: object
    properties:
      err:
        description: error details
        $ref: '#/definitions/ExceptionJSONInfoDTO'
      status:
        type: boolean
        description: status
      resTime:
        type: string
        description: response time
      response:
        type: array
        description: response
        items:
          $ref: '#/definitions/PreRegIdsByRegCenterIdResponseDTO'
    title: MainListResponseDTO«PreRegIdsByRegCenterIdResponseDTO»
  MainRequestDTO«CancelBookingDTO»:
    type: object
    properties:
      id:
        type: string
        description: request id
      ver:
        type: string
        description: request ver
      reqTime:
        type: string
        format: date-time
        description: request time
      request:
        description: request
        $ref: '#/definitions/CancelBookingDTO'
    title: MainRequestDTO«CancelBookingDTO»
  MainRequestDTO«PreRegIdsByRegCenterIdDTO»:
    type: object
    properties:
      id:
        type: string
        description: request id
      ver:
        type: string
        description: request ver
      reqTime:
        type: string
        format: date-time
        description: request time
      request:
        description: request
        $ref: '#/definitions/PreRegIdsByRegCenterIdDTO'
    title: MainRequestDTO«PreRegIdsByRegCenterIdDTO»
  MainResponseDTO«AvailabilityDto»:
    type: object
    properties:
      err:
        description: Error Details
        $ref: '#/definitions/ExceptionJSONInfoDTO'
      status:
        type: boolean
        description: Response Status
      resTime:
        type: string
        description: Response Time
      response:
        description: Response
        $ref: '#/definitions/AvailabilityDto'
    title: MainResponseDTO«AvailabilityDto»
  MainResponseDTO«BookingRegistrationDTO»:
    type: object
    properties:
      err:
        description: Error Details
        $ref: '#/definitions/ExceptionJSONInfoDTO'
      status:
        type: boolean
        description: Response Status
      resTime:
        type: string
        description: Response Time
      response:
        description: Response
        $ref: '#/definitions/BookingRegistrationDTO'
    title: MainResponseDTO«BookingRegistrationDTO»
  MainResponseDTO«CancelBookingResponseDTO»:
    type: object
    properties:
      err:
        description: Error Details
        $ref: '#/definitions/ExceptionJSONInfoDTO'
      status:
        type: boolean
        description: Response Status
      resTime:
        type: string
        description: Response Time
      response:
        description: Response
        $ref: '#/definitions/CancelBookingResponseDTO'
    title: MainResponseDTO«CancelBookingResponseDTO»
  MainResponseDTO«List«BookingStatusDTO»»:
    type: object
    properties:
      err:
        description: Error Details
        $ref: '#/definitions/ExceptionJSONInfoDTO'
      status:
        type: boolean
        description: Response Status
      resTime:
        type: string
        description: Response Time
      response:
        type: array
        description: Response
        items:
          $ref: '#/definitions/BookingStatusDTO'
    title: MainResponseDTO«List«BookingStatusDTO»»
  MainResponseDTO«string»:
    type: object
    properties:
      err:
        description: Error Details
        $ref: '#/definitions/ExceptionJSONInfoDTO'
      status:
        type: boolean
        description: Response Status
      resTime:
        type: string
        description: Response Time
      response:
        type: string
        description: Response
    title: MainResponseDTO«string»
  PreRegIdsByRegCenterIdDTO:
    type: object
    properties:
      registartion_center_id:
        type: string
        description: Registartion Center ID
      pre_registration_ids:
        type: array
        description: List of Pre-Registartion IDs
        items:
          type: string
    title: PreRegIdsByRegCenterIdDTO
  PreRegIdsByRegCenterIdResponseDTO:
    type: object
    properties:
      registartion_center_id:
        type: string
        description: Registartion Center ID
      pre_registration_ids:
        type: array
        description: List of Pre-Registartion IDs
        items:
          type: string
    title: PreRegIdsByRegCenterIdResponseDTO
  SlotDto:
    type: object
    properties:
      availability:
        type: integer
        format: int32
      fromTime:
        $ref: '#/definitions/LocalTime'
      toTime:
        $ref: '#/definitions/LocalTime'
    title: SlotDto